+++ src/custom/pages/Account/index.tsx
@@ -1,296 +1,54 @@
-import { useCallback, useEffect, useMemo, useState } from 'react'
-import {
-  ExtLink,
-  CardsWrapper,
-  Card,
-  CardActions,
-  BalanceDisplay,
-  ConvertWrapper,
-  VestingBreakdown,
-  CardsLoader,
-  CardsSpinner,
-} from 'pages/Account/styled'
-import { useActiveWeb3React } from 'hooks/web3'
-import { getBlockExplorerUrl } from 'utils'
-import { formatMax, formatSmartLocaleAware } from 'utils/format'
-import { MouseoverTooltipContent } from 'components/Tooltip'
-import { SupportedChainId as ChainId } from 'constants/chains'
-import AffiliateStatusCheck from 'components/AffiliateStatusCheck'
-import { HelpCircle } from 'components/Page'
-import { ButtonPrimary } from 'custom/components/Button'
-import vCOWImage from 'assets/cow-swap/vCOW.png'
-import SVG from 'react-inlinesvg'
-import ArrowIcon from 'assets/cow-swap/arrow.svg'
-import CowImage from 'assets/cow-swap/cow_v2.svg'
-import { useTokenBalance } from 'state/wallet/hooks'
-import { useVCowData, useSwapVCowCallback, useSetSwapVCowStatus, useSwapVCowStatus } from 'state/cowToken/hooks'
-import { V_COW_CONTRACT_ADDRESS, COW_CONTRACT_ADDRESS, AMOUNT_PRECISION } from 'constants/index'
-import { COW } from 'constants/tokens'
-import { useErrorModal } from 'hooks/useErrorMessageAndModal'
-import { OperationType } from 'components/TransactionConfirmationModal'
-import useTransactionConfirmationModal from 'hooks/useTransactionConfirmationModal'
-import AddToMetamask from 'components/AddToMetamask'
-import { Link } from 'react-router-dom'
-import CopyHelper from 'components/Copy'
-import { SwapVCowStatus } from 'state/cowToken/actions'
-import LockedGnoVesting from './LockedGnoVesting'
-import useBlockNumber from 'lib/hooks/useBlockNumber'
-import usePrevious from 'hooks/usePrevious'
-import { useCowFromLockedGnoBalances } from 'pages/Account/LockedGnoVesting/hooks'
-import { getProviderErrorMessage } from 'utils/misc'
-import WrapperWithNav from './WrapperWithNav'
+import { lazy, Suspense } from 'react'
+import { Switch, Route, useLocation } from 'react-router-dom'
+import { AccountMenu } from './Menu'
+import { Wrapper, AccountPageWrapper } from './Tokens/styled'
+import { Content } from 'components/Page'
+import { Routes } from 'constants/routes'
+import { Loading } from 'components/FlashingLoading'
 
-const COW_DECIMALS = COW[ChainId.MAINNET].decimals
+// Account pages
+const Balances = lazy(() => import(/* webpackChunkName: "account" */ 'pages/Account/Balances'))
+const TokensOverview = lazy(() => import(/* webpackChunkName: "tokens_overview" */ 'pages/Account/Tokens'))
+const Governance = lazy(() => import(/* webpackChunkName: "governance" */ 'pages/Account/Governance'))
+const Affiliate = lazy(() => import(/* webpackChunkName: "affiliate" */ 'pages/Account/Affiliate'))
+// Not found catch
+const NotFound = lazy(() => import(/* webpackChunkName: "affiliate" */ 'pages/error/NotFound'))
 
-// Number of blocks to wait before we re-enable the swap COW -> vCOW button after confirmation
-const BLOCKS_TO_WAIT = 2
-
-export default function Profile() {
-  const { account, chainId = ChainId.MAINNET, library } = useActiveWeb3React()
-  const previousAccount = usePrevious(account)
-
-  const blockNumber = useBlockNumber()
-  const [confirmationBlock, setConfirmationBlock] = useState<undefined | number>(undefined)
-  const [shouldUpdate, setShouldUpdate] = useState<boolean>(false)
-
-  const setSwapVCowStatus = useSetSwapVCowStatus()
-  const swapVCowStatus = useSwapVCowStatus()
-
-  // Locked GNO balance
-  const { loading: isLockedGnoLoading, ...lockedGnoBalances } = useCowFromLockedGnoBalances()
-
-  // Cow balance
-  const cow = useTokenBalance(account || undefined, chainId ? COW[chainId] : undefined)
-
-  // vCow balance values
-  const { unvested, vested, total, isLoading: isVCowLoading } = useVCowData()
-
-  // Boolean flags
-  const hasVestedBalance = vested && !vested.equalTo(0)
-  const hasVCowBalance = total && !total.equalTo(0)
-
-  const isSwapPending = swapVCowStatus === SwapVCowStatus.SUBMITTED
-  const isSwapInitial = swapVCowStatus === SwapVCowStatus.INITIAL
-  const isSwapConfirmed = swapVCowStatus === SwapVCowStatus.CONFIRMED
-  const isSwapDisabled = Boolean(
-    !hasVestedBalance || !isSwapInitial || isSwapPending || isSwapConfirmed || shouldUpdate
-  )
-
-  const isCardsLoading = useMemo(() => {
-    let output = isVCowLoading || isLockedGnoLoading || !library
-
-    // remove loader after 5 sec in any case
-    setTimeout(() => {
-      output = false
-    }, 5000)
-
-    return output
-  }, [isLockedGnoLoading, isVCowLoading, library])
-
-  const cowBalance = formatSmartLocaleAware(cow, AMOUNT_PRECISION) || '0'
-  const cowBalanceMax = formatMax(cow, COW_DECIMALS) || '0'
-  const vCowBalanceVested = formatSmartLocaleAware(shouldUpdate ? undefined : vested, AMOUNT_PRECISION) || '0'
-  const vCowBalanceVestedMax = vested ? formatMax(shouldUpdate ? undefined : vested, COW_DECIMALS) : '0'
-  const vCowBalanceUnvested = formatSmartLocaleAware(unvested, AMOUNT_PRECISION) || '0'
-  const vCowBalance = formatSmartLocaleAware(total, AMOUNT_PRECISION) || '0'
-  const vCowBalanceMax = total ? formatMax(total, COW_DECIMALS) : '0'
-
-  // Init modal hooks
-  const { handleSetError, handleCloseError, ErrorModal } = useErrorModal()
-  const { TransactionConfirmationModal, openModal, closeModal } = useTransactionConfirmationModal(
-    OperationType.CONVERT_VCOW
-  )
-
-  // Handle swaping
-  const { swapCallback } = useSwapVCowCallback({
-    openModal,
-    closeModal,
-  })
-
-  const handleVCowSwap = useCallback(async () => {
-    handleCloseError()
-
-    if (!swapCallback) {
-      return
-    }
-
-    setSwapVCowStatus(SwapVCowStatus.ATTEMPTING)
-
-    swapCallback()
-      .then(() => {
-        setSwapVCowStatus(SwapVCowStatus.SUBMITTED)
-      })
-      .catch((error) => {
-        console.error('[Profile::index::swapVCowCallback]::error', error)
-        setSwapVCowStatus(SwapVCowStatus.INITIAL)
-        handleSetError(getProviderErrorMessage(error))
-      })
-  }, [handleCloseError, handleSetError, setSwapVCowStatus, swapCallback])
-
-  const tooltipText = {
-    balanceBreakdown: (
-      <VestingBreakdown>
-        <span>
-          <i>Unvested</i> <p>{vCowBalanceUnvested} vCOW</p>
-        </span>
-        <span>
-          <i>Vested</i> <p>{vCowBalanceVested} vCOW</p>
-        </span>
-      </VestingBreakdown>
-    ),
-    vested: (
-      <div>
-        <p>
-          <strong>Vested vCOW</strong> is the portion of your vCOW token balance, which is fully available to convert to
-          COW token.
-        </p>
-        <p>
-          This includes any vCOW received through an <strong>airdrop.</strong>
-        </p>
-        <p>When converting your vested vCOW balance to COW, your entire vested balance will be converted.</p>
-      </div>
-    ),
+function _getPropsFromRoute(route: string) {
+  switch (route) {
+    case Routes.ACCOUNT:
+      return ['account-general', 'General']
+    case Routes.ACCOUNT_AFFILIATE:
+      return ['account-affiliate', 'Affiliate']
+    case Routes.ACCOUNT_GOVERNANCE:
+      return ['account-governance', 'Governance']
+    case Routes.ACCOUNT_TOKENS:
+      return ['account-tokens', 'Tokens']
+    default:
+      return []
   }
+}
 
-  const renderConvertToCowContent = useCallback(() => {
-    let content = null
-
-    if (isSwapPending) {
-      content = <span>Converting vCOW...</span>
-    } else if (isSwapConfirmed) {
-      content = <span>Successfully converted!</span>
-    } else {
-      content = (
-        <>
-          Convert to COW <SVG src={ArrowIcon} />
-        </>
-      )
-    }
-
-    return content
-  }, [isSwapConfirmed, isSwapPending])
-
-  // Fixes the issue with change in status after swap confirmation
-  // Makes sure to wait 2 blocks after confirmation to enable the swap button again
-  useEffect(() => {
-    if (isSwapConfirmed && !confirmationBlock) {
-      setConfirmationBlock(blockNumber)
-      setShouldUpdate(true)
-    }
-
-    if (!confirmationBlock || !blockNumber) {
-      return
-    }
-
-    if (isSwapConfirmed && blockNumber - confirmationBlock > BLOCKS_TO_WAIT && hasVestedBalance) {
-      setSwapVCowStatus(SwapVCowStatus.INITIAL)
-      setConfirmationBlock(undefined)
-      setShouldUpdate(false)
-    }
-  }, [blockNumber, confirmationBlock, hasVestedBalance, isSwapConfirmed, setSwapVCowStatus, shouldUpdate])
-
-  // Reset swap button status on account change
-  useEffect(() => {
-    if (account && previousAccount && account !== previousAccount && !isSwapInitial) {
-      setSwapVCowStatus(SwapVCowStatus.INITIAL)
-    }
-  }, [account, isSwapInitial, previousAccount, setSwapVCowStatus])
-
-  const currencyCOW = COW[chainId]
-
+export default function Account() {
+  const { pathname } = useLocation()
+  const [id, name] = _getPropsFromRoute(pathname)
   return (
-    <WrapperWithNav name="General" id="account-general">
-      <TransactionConfirmationModal />
-      <ErrorModal />
-
-      {chainId && chainId === ChainId.MAINNET && <AffiliateStatusCheck />}
-
-      <CardsWrapper>
-        {isCardsLoading ? (
-          <CardsWrapper>
-            <CardsLoader>
-              <CardsSpinner size="24px" />
-            </CardsLoader>
-          </CardsWrapper>
-        ) : (
-          <>
-            {hasVCowBalance && (
-              <Card showLoader={isVCowLoading || isSwapPending}>
-                <BalanceDisplay hAlign="left">
-                  <img src={vCOWImage} alt="vCOW token" width="56" height="56" />
-                  <span>
-                    <i>Total vCOW balance</i>
-                    <b>
-                      <span title={`${vCowBalanceMax} vCOW`}>{vCowBalance} vCOW</span>{' '}
-                      <MouseoverTooltipContent content={tooltipText.balanceBreakdown} wrap>
-                        <HelpCircle size={14} />
-                      </MouseoverTooltipContent>
-                    </b>
-                  </span>
-                </BalanceDisplay>
-                <ConvertWrapper>
-                  <BalanceDisplay titleSize={18} altColor={true}>
-                    <i>
-                      Vested{' '}
-                      <MouseoverTooltipContent content={tooltipText.vested} wrap>
-                        <HelpCircle size={14} />
-                      </MouseoverTooltipContent>
-                    </i>
-                    <b title={`${vCowBalanceVestedMax} vCOW`}>{vCowBalanceVested}</b>
-                  </BalanceDisplay>
-                  <ButtonPrimary onClick={handleVCowSwap} disabled={isSwapDisabled}>
-                    {renderConvertToCowContent()}
-                  </ButtonPrimary>
-                </ConvertWrapper>
-
-                <CardActions>
-                  <ExtLink href={getBlockExplorerUrl(chainId, V_COW_CONTRACT_ADDRESS[chainId], 'token')}>
-                    View contract ↗
-                  </ExtLink>
-                  <CopyHelper toCopy={V_COW_CONTRACT_ADDRESS[chainId]}>
-                    <div title="Click to copy token contract address">Copy contract</div>
-                  </CopyHelper>
-                </CardActions>
-              </Card>
-            )}
-
-            <Card>
-              <BalanceDisplay titleSize={26}>
-                <img src={CowImage} alt="Cow Balance" height="80" width="80" />
-                <span>
-                  <i>Available COW balance</i>
-                  <b title={`${cowBalanceMax} COW`}>{cowBalance} COW</b>
-                </span>
-              </BalanceDisplay>
-              <CardActions>
-                <ExtLink
-                  title="View contract"
-                  href={getBlockExplorerUrl(chainId, COW_CONTRACT_ADDRESS[chainId], 'token')}
-                >
-                  View contract ↗
-                </ExtLink>
-
-                {library?.provider?.isMetaMask && <AddToMetamask shortLabel currency={currencyCOW} />}
-
-                {!library?.provider?.isMetaMask && (
-                  <CopyHelper toCopy={COW_CONTRACT_ADDRESS[chainId]}>
-                    <div title="Click to copy token contract address">Copy contract</div>
-                  </CopyHelper>
-                )}
-
-                <Link to={`/swap?outputCurrency=${COW_CONTRACT_ADDRESS[chainId]}`}>Buy COW</Link>
-              </CardActions>
-            </Card>
-
-            <LockedGnoVesting
-              {...lockedGnoBalances}
-              loading={isLockedGnoLoading}
-              openModal={openModal}
-              closeModal={closeModal}
-            />
-          </>
-        )}
-      </CardsWrapper>
-    </WrapperWithNav>
+    <Wrapper>
+      <AccountMenu />
+      <Suspense fallback={Loading}>
+        <AccountPageWrapper>
+          <Content>
+            <h2 id={id}>{name}</h2>
+            <Switch>
+              <Route exact path={Routes.ACCOUNT} component={Balances} />
+              <Route exact strict path={Routes.ACCOUNT_GOVERNANCE} component={Governance} />
+              <Route exact strict path={Routes.ACCOUNT_TOKENS} component={TokensOverview} />
+              <Route exact strict path={Routes.ACCOUNT_AFFILIATE} component={Affiliate} />
+              <Route component={NotFound} />
+            </Switch>
+          </Content>
+        </AccountPageWrapper>
+      </Suspense>
+    </Wrapper>
   )
 }
